pyproject.toml
[project]
name = "ball-compliance-regulatory-agent"
version = "0.1.0"
description = "Enterprise Regulatory AI"
authors = [
    {name = "Anurag Sake"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "qdrant-client (>=1.16.2,<2.0.0)",
    "openai (>=2.15.0,<3.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
